requirements.txt
fastapi
pyjwt
uvicorn
requests